+++ src/main/java/com/nhnacademy/booklay/server/controller/admin/product/AuthorAdminController.java
@@ -18,6 +18,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+
+/**
+ * @author 최규태
+ */
+
 @Slf4j
 @RestController
 @RequestMapping("/admin/author")
