+++ src/main/java/com/nhnacademy/booklay/server/controller/admin/product/TagAdminController.java
@@ -21,6 +21,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+ * @author 최규태
+ */
+
 @Slf4j
 @RestController
 @RequestMapping("/admin/tag")
