+++ src/main/java/com/nhnacademy/booklay/server/service/product/AuthorService.java
@@ -8,6 +8,10 @@ import com.nhnacademy.booklay.server.entity.Author;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+/**
+ * @author 최규태
+ */
+
 public interface AuthorService{
   void createAuthor(CreateAuthorRequest author);
