+++ src/main/java/com/nhnacademy/booklay/server/service/product/ProductAuthorService.java
@@ -2,6 +2,10 @@ package com.nhnacademy.booklay.server.service.product;
 
 import com.nhnacademy.booklay.server.entity.ProductAuthor;
 
+/**
+ * @author 최규태
+ */
+
 public interface ProductAuthorService{
   ProductAuthor createProductAuthor(ProductAuthor productAuthor);
