--- src/main/java/com/nhnacademy/booklay/server/service/product/ProductTagService.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.booklay.server.service.product;
-
-public interface ProductTagService {
-
-}
