+++ src/main/java/com/nhnacademy/booklay/server/service/product/SubscribeService.java
@@ -2,6 +2,10 @@ package com.nhnacademy.booklay.server.service.product;
 
 import com.nhnacademy.booklay.server.entity.Subscribe;
 
+/**
+ * @author 최규태
+ */
+
 public interface SubscribeService {
   Subscribe createSubscribe(Subscribe subscribe);
