+++ src/main/java/com/nhnacademy/booklay/server/service/product/WishlistService.java
@@ -2,6 +2,10 @@ package com.nhnacademy.booklay.server.service.product;
 
 import com.nhnacademy.booklay.server.dto.product.request.CreateWishlistRequest;
 
+/**
+ * @author 최규태
+ */
+
 public interface WishlistService {
 
   void createWishlist(CreateWishlistRequest request);
